# swagger-spec.yaml
openapi: 3.0.0
info:
  title: API de Postagem Social
  version: 1.0.0
  description: API para postar conteúdo em diversas redes sociais.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
...
    SocialPostRequest:
      type: object
      properties:
        text:
          type: string
        images:
          type: array
          items:
            type: string
            format: byte
        tags:
          type: array
          items:
            type: string
paths:
  /auth/login:
    post:
      summary: Autentica o usuário e retorna um token JWT.
      tags:
        - Autenticação
      description: >
        Autentica na API utilizando um conjunto de credenciais pré-definidas e
        retorna um token de acesso JWT. Este token é necessário para autorizar
        todas as outras requisições aos endpoints protegidos da aplicação. O
        token gerado possui uma validade de 24 horas.


        **Corpo da Requisição:**

        * **`username`** (string, obrigatório): O nome de usuário definido para a API.

        * **`password`** (string, obrigatório): A senha definida para a API.

        * **`accessToken`** (string, obrigatório): Um token de acesso estático que atua como uma camada adicional de segurança.


        **Corpo da Resposta:**

        * Retorna um objeto contendo a mensagem de sucesso, o token JWT e a data de expiração no formato ISO 8601 (status `200 OK`).

        * Retorna um erro `401 Unauthorized` caso qualquer uma das credenciais fornecidas seja inválida.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/SocialPostRequest"
                - type: object
                  properties:
                    accessToken:
                      type: string
            example:
              username: seu-usuario
              password: sua-senha-forte
              accessToken: um-token-secreto-e-longo
      responses:
        "200":
          description: Login bem-sucedido.
          content:
            application/json:
              example:
                message: Autenticação bem-sucedida!
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                expiresAt: 2025-09-14T22:38:59.123Z
        "401":
          description: Credenciais inválidas.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/token/refresh:
    post:
      summary: Renova um token JWT expirado.
      tags:
        - Autenticação
      description: >
        Gera um novo token de acesso JWT com validade de 24 horas a partir de um
        token existente, mesmo que este já esteja expirado. Este endpoint é
        utilizado para manter a sessão do cliente ativa sem a necessidade de
        enviar as credenciais de `username` e `password` novamente.


        **Cabeçalho da Requisição:**

        * **`Authorization`**: Deve conter o token JWT atual (mesmo que expirado) no formato `Bearer SEU_TOKEN_AQUI`.


        **Corpo da Requisição:**

        * **`accessToken`** (string, obrigatório): O mesmo token de acesso estático usado no login, para validar a legitimidade da requisição de renovação.


        **Corpo da Resposta:**

        * Retorna um objeto com a mensagem de sucesso, o **novo** token JWT e a sua **nova** data de expiração (status `200 OK`).

        * Retorna um erro `401 Unauthorized` se o token antigo for inválido (assinatura incorreta) ou se o `accessToken` não corresponder.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/SocialPostRequest"
                - type: object
                  properties:
                    accessToken:
                      type: string
            example:
              accessToken: um-token-secreto-e-longo
      responses:
        "200":
          description: Token renovado com sucesso.
        "401":
          description: Token inválido/ausente.
  /bluesky/post:
    post:
      summary: Cria um novo post (skeet) no Bluesky.
      tags:
        - Bluesky
      description: >
        Publica um novo post (skeet) na conta do Bluesky autenticada. A postagem
        pode conter texto e até 4 imagens.


        **Corpo da Requisição:**

        * **`text`** (string, obrigatório): O conteúdo principal do post.

        * **`images`** (array, opcional): Uma lista de até 4 imagens no formato Data URL (base64). Imagens maiores que o limite da plataforma (~976KB) serão automaticamente otimizadas.

        * **`tags`** (array, opcional): Uma lista de tags que serão convertidas em hashtags e adicionadas ao final do texto.


        **Corpo da Resposta:**

        * Retorna um objeto com os dados do post criado com sucesso (status `201 Created`).

        * Retorna um erro `400 Bad Request` se os dados forem inválidos (ex: mais de 4 imagens).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/SocialPostRequest"
                - type: object
                  properties:
                    text:
                      type: string
                    images:
                      type: array
                      items:
                        type: base64
                    tags:
                      type: array
                      items:
                        type: string
            example:
              text: Este é um tweet de exemplo!
              tags:
                - api
                - teste
              images:
                - data:image/png;base64,iVBORw0KGgo...
      responses:
        "201":
          description: Post criado com sucesso.
...